kstd::arch

Module x86

Platform-specific code for the x86 platform.

Modules

boot

The x86 boot module defines the entrypoints of Astros and the corresponding headers for different x86 boot protocols.

device

Device-related APIs. This module mainly contains the APIs that should exposed to the device driver like PCI, RTC

iommu

The IOMMU support.

qemu

Provides the ability to exit QEMU and return a value as debug result.

serial

The console I/O.

task

The architecture support of context switch.

timer

The timer support.

trap

Handles trap.

Functions

read_random

Reads a hardware generated 64-bit random value.

read_tsc

Reads the current value of the processor’s time-stamp counter (TSC).

tsc_freq

Returns the frequency of TSC. The unit is Hz.