kstd::arch::x86::device

Module io_port

I/O port access.

Structs

IoPort

An I/O port, representing a specific address in the I/O address of x86.

ReadOnlyAccess

An access marker type indicating that a port is only allowed to read values.

ReadWriteAccess

An access marker type indicating that a port is allowed to read or write values.

WriteOnlyAccess

An access marker type indicating that a port is only allowed to write values.

Traits

IoPortReadAccess

A marker trait for access types which allow reading port values.

IoPortWriteAccess

A marker trait for access types which allow writing port values.

PortRead

A helper trait that implements the read port operation.

PortWrite

A helper trait that implements the write port operation.