kstd::arch::x86

Module serial

The console I/O.

Functions

print

Prints the formatted arguments to the standard output using the serial port.

register_console_input_callback

Registers a callback function to be called when there is console input.

send

Sends a byte on the serial port.

Type Aliases

InputCallback

The callback function for console input.