kstd::collections

Module xarray

This module introduces the xarray crate and provides relevant support and interfaces for XArray.

Structs

Cursor

A Cursor can traverse in the XArray by setting or increasing the target index and can perform read-only operations to the target item represented by the target index.

CursorMut

A CursorMut can traverse in the XArray by setting or increasing the target index and can perform read-write operations to the target item represented by the target index.

XArray

XArray is an abstract data type functioning like an expansive array of items where each item must be an 8-byte object, such as Arc<T> or Box<T>.

Enums

XMark

The mark type used in the XArray.