kstd

Module mm

Virtual memory (VM).

Re-exports

pub use self::frame::allocator::FrameAllocOptions;

pub use self::frame::segment::Segment;

pub use self::frame::segment::USegment;

pub use self::frame::unique::UniqueFrame;

pub use self::frame::untyped::AnyUFrameMeta;

pub use self::frame::untyped::UFrame;

pub use self::frame::untyped::UntypedMem;

pub use self::frame::Frame;

pub use self::vm_space::VmSpace;

Modules

frame

Frame (physical memory page) management.

heap

Manages the kernel heap using slab or buddy allocation strategies.

tlb

TLB flush operations.

vm_space

Virtual memory space management.

Structs

DmaCoherent

A coherent (or consistent) DMA mapping, which guarantees that the device and the CPU can access the data in parallel.

DmaStream

A streaming DMA mapping. Users must synchronize data before reading or after writing to ensure consistency.

DmaStreamSlice

A slice of streaming DMA mapping.

PageFlags

Page protection permissions and access status.

PageProperty

The property of a mapped virtual memory page.

VmReader

VmReader is a reader for reading data from a contiguous range of memory.

VmWriter

VmWriter is a writer for writing data to a contiguous range of memory.

Enums

CachePolicy

A type to control the cacheability of the main memory.

DmaDirection

DmaDirection limits the data flow direction of DmaStream and prevents users from reading and writing to DmaStream unexpectedly.

Fallible

A marker type used for VmReader and VmWriter, representing whether reads or writes on the underlying memory region are fallible.

Infallible

A marker type used for VmReader and VmWriter, representing whether reads or writes on the underlying memory region are infallible.

Constants

KERNEL_VADDR_RANGE

The kernel address space.

MAX_USERSPACE_VADDR

The maximum virtual address of user space (non inclusive).

PAGE_SIZE

The page size

Traits

FallibleVmRead

Fallible memory read from a VmWriter.

FallibleVmWrite

Fallible memory write from a VmReader.

HasDaddr

A trait for types that have mapped address in the device address space.

HasPaddr

Gets physical address trait

PodOnce

A marker trait for POD types that can be read or written with one instruction.

VmIo

A trait that enables reading/writing data from/to a VM object, e.g., USegment, Vec<UFrame> and UFrame.

VmIoOnce

A trait that enables reading/writing data from/to a VM object using one non-tearing memory load/store.

Functions

is_page_aligned

Checks if the given address is page-aligned.

Type Aliases

Daddr

The device address.

Paddr

Physical addresses.

PagingLevel

The level of a page table node or a frame.

Vaddr

Virtual addresses.