kstd::mm::frame::untyped

Type Alias UFrame

pub type UFrame = Frame<dyn AnyUFrameMeta>;

A smart pointer to an untyped frame with any metadata.

The metadata of the frame is not known at compile time but the frame must be an untyped one. An UFrame as a parameter accepts any type of untyped frame metadata.

The usage of this frame will not be changed while this object is alive.

Aliased Type

struct UFrame { /* private fields */ }