kstd::sync

Trait OwnerPtr

pub unsafe trait OwnerPtr: Send + 'static {
    type Target;

    // Required methods
    fn new(value: Self::Target) -> Self;
    fn into_raw(self) -> NonNull<Self::Target>;
    unsafe fn from_raw(ptr: NonNull<Self::Target>) -> Self;
}

A trait that abstracts pointers that have the ownership of the objects they refer to.

The most typical examples smart pointer types like Box<T> and Arc<T>, which can be converted to and from the raw pointer type of *const T.

Safety

This trait must be implemented correctly (according to the doc comments for each method). Types like Rcu rely on this assumption to safely use the raw pointers.

Required Associated Types

type Target

The target type that this pointer refers to.

Required Methods

fn new(value: Self::Target) -> Self

Creates a new pointer with the given value.

fn into_raw(self) -> NonNull<Self::Target>

Converts to a raw pointer.

Each call to into_raw must be paired with a call to from_raw in order to avoid memory leakage.

The resulting raw pointer must be valid to be immutably accessed or borrowed until from_raw is called.

unsafe fn from_raw(ptr: NonNull<Self::Target>) -> Self

Converts back from a raw pointer.

Safety

The raw pointer must have been previously returned by a call to into_raw.

The raw pointer must not be used after calling from_raw.

Note that the second point is a hard requirement: Even if the resulting value has not (yet) been dropped, the pointer cannot be used because it may break Rust aliasing rules (e.g., Box<T> requires the pointer to be unique and thus never aliased).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.