kstd::task::atomic_mode

Trait InAtomicMode

pub unsafe trait InAtomicMode { }

A marker trait for guard types that enforce the atomic mode.

Key kernel primitives such as SpinLock and Rcu rely on the atomic mode for correctness or soundness. The existence of such a guard guarantees that the current task is executing in the atomic mode.

Safety

The implementer must ensure that the atomic mode is maintained while the guard type is alive.